requirements-dev.txt
# 개발 환경 의존성 (Python 3.13 호환)
-r requirements.txt

# 테스팅
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-flask>=1.3.0,<2.0.0

# 코드 품질
black>=23.7.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
isort>=5.12.0,<6.0.0

# 개발 도구
python-dotenv>=1.0.0,<2.0.0
watchdog>=3.0.0,<4.0.0
